Consolidate the fourth script in the pipeline
